Replace dot with dash in slugs
The `.` character is allowed in slugs, but Jekyll's array reference into data, `site.data.comments[slug]`, doesn't like it if the `slug` has a `.` character in it. So this replaces the dot with a dash. I'll have to make a corresponding change to https://github.com/damieng/jekyll-blog-comments-azure/blob/master/JekyllBlogCommentsAzure/PostCommentToPullRequestFunction.cs